migrating react to react-vite
