+++ src/components/SearchResult.jsx
